+++ public/run-model.sh
@@ -2,7 +2,7 @@
 
 
 #this is the version of the compiled llama.cpp core, This is generally requires to support the new launched models.
-VERSION="b3658" # Change this if the version changes
+VERSION="b3899" # Change this if the version changes
 INSTALL_DIR="$HOME/.pyano"
 #This is where the compiled version of llama.cpp will be unzeipped that has llama-server binary to run llama.cpp server.
 BUILD_DIR="$INSTALL_DIR/build/bin"
@@ -23,37 +23,48 @@ MMAP="${MMAP:-false}"
 
 # Function to download and unzip if the version is not present
 download_and_unzip() {
-    # Check if curl or wget is installed and set DOWNLOAD_CMD accordingly
-    if command -v curl &> /dev/null; then
-        DOWNLOAD_CMD="curl -Lo"
-    elif command -v wget &> /dev/null; then
-        DOWNLOAD_CMD="wget -P"
-    else
-        echo "Neither curl nor wget is installed. Installing curl..."
-        if [[ "$OSTYPE" == "linux-gnu"* ]]; then
-            sudo apt-get update && sudo apt-get install -y curl
-        elif [[ "$OSTYPE" == "darwin"* ]]; then
-            brew install curl
+    # Path to the llama-server binary
+    Llama_Server_Path="$INSTALL_DIR/build/bin/llama-server"
+    echo "checking for $Llama_Server_Path"
+
+    # Check if llama-server binary exists
+    if [[ ! -f "$Llama_Server_Path" ]]; then
+        echo "llama-server not found. Downloading and unzipping the new version..."
+
+        # Check if curl or wget is installed and set DOWNLOAD_CMD accordingly
+        if command -v curl &> /dev/null; then
+            DOWNLOAD_CMD="curl -Lo"
+        elif command -v wget &> /dev/null; then
+            DOWNLOAD_CMD="wget -P"
         else
-            echo "Unsupported OS for automatic curl installation. Please install curl or wget manually."
-            exit 1
+            echo "Neither curl nor wget is installed. Installing curl..."
+            if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+                sudo apt-get update && sudo apt-get install -y curl
+            elif [[ "$OSTYPE" == "darwin"* ]]; then
+                brew install curl
+            else
+                echo "Unsupported OS for automatic curl installation. Please install curl or wget manually."
+                exit 1
+            fi
+            DOWNLOAD_CMD="curl -Lo"
         fi
-        DOWNLOAD_CMD="curl -Lo"
-    fi
 
-    # Create the ~/.pyano/ directory if it doesn't exist
-    mkdir -p $INSTALL_DIR
+        # Create the ~/.pyano/ directory if it doesn't exist
+        mkdir -p $INSTALL_DIR
 
-    # Download the appropriate file based on the OS
-    if [[ ! -f "$INSTALL_DIR/$ZIP_FILE" ]]; then
-        echo "Downloading $ZIP_FILE..."
-        $DOWNLOAD_CMD $INSTALL_DIR/$ZIP_FILE $DOWNLOAD_URL
+        # Download the appropriate file based on the OS
+        if [[ ! -f "$INSTALL_DIR/$ZIP_FILE" ]]; then
+            echo "Downloading $ZIP_FILE for llama.cpp ..."
+            $DOWNLOAD_CMD $INSTALL_DIR/$ZIP_FILE $DOWNLOAD_URL
 
-        # Unzip the downloaded file
-        echo "Unzipping $ZIP_FILE..."
-        unzip $INSTALL_DIR/$ZIP_FILE -d $INSTALL_DIR/
+            # Unzip the downloaded file
+            echo "Unzipping $ZIP_FILE..."
+            unzip $INSTALL_DIR/$ZIP_FILE -d $INSTALL_DIR/
+        else
+            echo "$ZIP_FILE already exists, skipping download and unzip."
+        fi
     else
-        echo "$ZIP_FILE already exists, skipping download and unzip."
+        echo "llama-server already exists at $Llama_Server_Path, skipping download."
     fi
 }
 
@@ -98,17 +109,7 @@ check_and_download_model() {
     fi
 }
 
-# Ensure MODEL_PATH is set
-MODEL_PATH="$MODEL_DIR/$MODEL_NAME"
 
-check_and_download_model
-
-# Set download info based on the OS
-set_download_info
-
-# Download and unzip the file
-
-install_requirements_llama
 
 # Ensure MODEL_PATH is set
 if [ -z "$MODEL_PATH" ]; then
@@ -117,7 +118,7 @@ if [ -z "$MODEL_PATH" ]; then
 fi
 
 # Download and unzip if necessary
-download_and_unzip
+
 
 # Calculate the number of CPU cores
 get_num_cores() {
@@ -138,29 +139,35 @@ get_num_cores() {
         return 1
     fi
 }
+
+check_and_download_model
+
+# Set download info based on the OS
+set_download_info
+
+# Download and unzip the file
+
+install_requirements_llama
+download_and_unzip
 get_num_cores
 echo "Model being used $MODEL_PATH"
 echo "Number of cores are  $num_cores"
 
-# Run the server command with parameters from the config
+# Run the server command
 $BUILD_DIR/llama-server \
-    -m "$MODEL_PATH" \
-    --ctx-size "$CTX" \
-    --parallel 2 \
-    --n-gpu-layers "$GPU_LAYERS_OFFLOADED" \
-    --port 52555 \
-    --threads "$num_cores" \
-    --color \
-    --metrics \
-    --batch-size "$BATCH_SIZE" \
-    $( [ "$MMAP" == "true" ]|| echo "--no-mmap" ) \
-    $( [ "$MLOCK" == "true" ] && echo "--mlock" ) \
-    --conversation \
+  -m $MODEL_PATH \
+  --ctx-size $CTX \
+  --parallel 2 \
+  --n-gpu-layers $GPU_LAYERS_OFFLOADED\
+  --port 52555 \
+  --threads $num_cores \
+  --metrics \
+    --batch-size $BATCH_SIZE \
+    --no-mmap \
     --flash-attn \
-    --cache-type-k f16 \
-    --cache-type-v f16 \
-    --prompt-cache-all \
-    --repeat-last-n 64 \
-    --repeat-penalty 1.3 \
-    --top-k 40 \
-    --top-p 0.9
+  --cache-type-k f16 \
+  --cache-type-v f16 \
+   --repeat-last-n 64 \
+   --repeat-penalty 1.3 \
+   --top-k 40 \
+   --top-p 0.9
